Изменил страницу со списком товаров и создал страницу для добавления и изменения товаров на сиранице со списком(пока не совсем работает)
