--- Pages/AddProductPage.cshtml
@@ -1,4 +1,0 @@
-﻿@page
-@model Склад.Pages.AddProductPageModel
-@{
-}
