--- Pages/AddProductPage.cshtml.cs
@@ -1,12 +1,0 @@
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace Склад.Pages
-{
-    public class AddProductPageModel : PageModel
-    {
-        public void OnGet()
-        {
-        }
-    }
-}
