--- Pages/EditPage.cshtml
@@ -1,26 +1,0 @@
-﻿@page
-@model Склад.Pages.EditPageModel
-@{
-    ViewData["Title"] = "Редактировать товар";
-}
-
-<h1>Редактировать товар</h1>
-
-<form method="post">
-    <div class="form-group">
-        <label asp-for="Product.Name"></label>
-        <input asp-for="Product.Name" class="form-control" />
-    </div>
-    <div class="form-group">
-        <label asp-for="Product.Price"></label>
-        <input asp-for="Product.Price" class="form-control" />
-    </div>
-    <div class="form-group">
-        <label asp-for="Product.Quantity"></label>
-        <input asp-for="Product.Quantity" class="form-control" />
-    </div>
-    <input type="hidden" asp-for="Product.ProductId" />
-    <button type="submit" class="btn btn-primary">Сохранить</button>
-</form>
-
-<a asp-page="./ProductsList">Назад к списку</a>
