+++ Pages/ProductsList.cshtml
@@ -2,26 +2,31 @@
 @model Склад.Pages.ProductsListModel
 @{
 }
-<h2>Товары</h2>
-    <table>
-        <thead>
-            <tr>
-                <th>Id</th>
-                <th>Наименование</th>
-                <th>Цена</th>
-                <th>Количество</th>
-            </tr>
-        </thead>
-        <tbody>
-            @foreach (var product in Model.Products)
-            {
-                <tr>
-                    <td>@product.ProductId</td>
-                    <td>@product.Name</td>
-                    <td>@product.Price</td>
-                    <td>@product.Quantity</td>
-                <a asp-page="./Edit" asp-route-id="@product.ProductId">Редактировать</a>
-                </tr>
-            }
-    </tbody>
-</table>
+
+<h2>Products</h2>
+
+<p>
+    <a class="btn btn-primary" href="@Url.Page("Product", new { id = 0 })">Добавить новый товар</a>
+</p>
+
+<div class="row row-cols-1 row-cols-md-3 g-4">
+    @foreach (var product in Model.Products)
+    {
+        <div class="col">
+            <div class="card h-100">
+                <div class="card-body">
+                    <h5 class="card-title">@product.Name</h5>
+                    <p class="card-text">Грамм: @product.Price</p>
+                    <p class="card-text">Грамм: @product.Quantity</p>
+                    <p class="card-text">Цена: @product.CategoryId</p>
+                    <p class="card-text">Цена: @product.SupplierId</p>
+                   
+                </div>
+                <div class="card-footer d-flex justify-content-between">
+                    <a href="@Url.Page("Product", new { id = product.ProductId })" class="btn btn-warning">Изменить</a>
+                    <a href="javascript:void(0);" class="btn btn-danger" onclick="if (confirm('Are you sure you want to delete?')) { window.location.href = '@Url.Page("ProductDelete", new { id = product.ProductId })'; }">Удалить</a>
+                </div>
+            </div>
+        </div>
+    }
+</div>
